+++ prompt.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 import argparse
 import datetime
@@ -923,6 +923,7 @@ class PromptInterface(object):
 def main():
     parser = argparse.ArgumentParser()
 
+    # Network group
     group = parser.add_mutually_exclusive_group()
     group.add_argument("-m", "--mainnet", action="store_true", default=False,
                        help="Use MainNet instead of the default TestNet")
@@ -932,9 +933,16 @@ def main():
                        help="Use the CoZ network instead of the default TestNet")
     group.add_argument("-c", "--config", action="store", help="Use a specific config file")
 
+    # Theme
     parser.add_argument("-t", "--set-default-theme", dest="theme",
                         choices=["dark", "light"],
                         help="Set the default theme to be loaded from the config file. Default: 'dark'")
+
+    # Verbose
+    parser.add_argument("-v", "--verbose", action="store_true", default=False,
+                        help="Show smart-contract events by default")
+
+    # Show the neo-python version
     parser.add_argument("--version", action="version",
                         version="neo-python v{version}".format(version=__version__))
 
@@ -952,6 +960,9 @@ def main():
 
     if args.theme:
         preferences.set_theme(args.theme)
+
+    if args.verbose:
+        settings.set_log_smart_contract_events(True)
 
     # Instantiate the blockchain and subscribe to notifications
     blockchain = LevelDBBlockchain(settings.LEVELDB_PATH)