+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ All notable changes to this project are documented in this file.
 - update to neo-boa==0.3.7
 - `api-server.py <https://github.com/CityOfZion/neo-python/blob/development/api-server.py>`_: Improved logging setup. See the options with ``./api-server.py -h``
 - Renamed ``neo.api.REST.NotificationRestApi`` to ``neo.api.REST.RestApi``
+- Added ``-v/--verbose`` argument to prompt.py, which makes prompt.py show smart contract events by default
+
 
 [0.5.3] 2018-03-04
 ------------------